bot/backend/requirements.txt
pyautogen
flask
spacy
python-dotenv
autogen
fastapi
uvicorn
pydantic